+++ cmd/root.go
@@ -19,6 +19,7 @@ func SetBuildInfo(version, commit, date string) {
 	buildVersion = version
 	buildCommit = commit
 	buildDate = date
+	rootCmd.Version = version
 }
 
 var rootCmd = &cobra.Command{